+++ spec/components/credit_card_spec.rb
@@ -16,6 +16,14 @@ describe Minfraud::Components::CreditCard do
       end.to raise_exception(Minfraud::InvalidInputError)
     end
 
+    it 'raises an exception for an invalid last_digits' do
+      expect do
+        Minfraud::Components::CreditCard.new(
+          last_digits: '6',
+        )
+      end.to raise_exception(Minfraud::InvalidInputError)
+    end
+
     it 'raises an exception for an invalid last_4_digits' do
       expect do
         Minfraud::Components::CreditCard.new(
@@ -40,19 +48,41 @@ describe Minfraud::Components::CreditCard do
       end.to raise_exception(Minfraud::InvalidInputError)
     end
 
-    it 'does not raise an exception for valid values' do
-      Minfraud::Components::CreditCard.new(
+    it 'does not raise an exception for valid values (deprecated last_4_digits)' do
+      @cc = Minfraud::Components::CreditCard.new(
         issuer_id_number:         '123456',
+        last_4_digits:            '1234',
+        token:                    'abcd',
+        was_3d_secure_successful: true,
+      )
+      expect(@cc.last_digits).to be(@cc.last_4_digits)
+      expect(@cc.last_digits).to be('1234')
+    end
+
+    it 'does not raise an exception for valid values (eight digit issuer_id_number)' do
+      Minfraud::Components::CreditCard.new(
+        issuer_id_number:         '12345678',
         last_4_digits:            '1234',
         token:                    'abcd',
         was_3d_secure_successful: true,
       )
     end
 
+    it 'does not raise an exception for valid values (two digit last_digits)' do
+      @cc = Minfraud::Components::CreditCard.new(
+        issuer_id_number:         '12345678',
+        last_digits:              '34',
+        token:                    'abcd',
+        was_3d_secure_successful: true,
+      )
+      expect(@cc.last_digits).to be(@cc.last_4_digits)
+      expect(@cc.last_digits).to be('34')
+    end
+
     it 'does not raise an exception for valid values (token is all digits)' do
       Minfraud::Components::CreditCard.new(
         issuer_id_number:         '123456',
-        last_4_digits:            '1234',
+        last_digits:              '1234',
         token:                    '1' * 20,
         was_3d_secure_successful: true,
       )